fix(lint): fixture linter
