+++ backend/tests/integration/api/test_jwks.py
@@ -1,26 +1,11 @@
 """Integration tests for the JWKS endpoint."""
-
 import json
 from unittest.mock import AsyncMock, patch
 
 import httpx
 import pytest
 
-from app.schemas.jwks import JWK, JWKSet
-
-
-@pytest.fixture
-def sample_jwks():
-    """Create a sample JWKSet for testing."""
-    jwk = JWK(
-        kty="RSA",
-        use="sig",
-        kid="test-key-1",
-        alg="RS256",
-        n="test-modulus",
-        e="AQAB",
-    )
-    return JWKSet(keys=[jwk])
+from tests.fixtures.deduplicated import sample_jwks
 
 
 @pytest.mark.asyncio
